Create test_channels.hpp
